% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\name{Atomic}
\alias{Atomic}
\title{Atomic bombs data}
\format{
A  \code{data.frame} containing 1039 observations on the following 5 variables:
\describe{
\item{\code{y.perc}}{the percentage of cells with chromosomal abnormalities.}
\item{\code{y}}{the number of cells with chromosomal abnormalities.}
\item{\code{n}}{the number of analyzed cells.  It is fixed to 100 for all the survivors.}
\item{\code{dose}}{a quantitative measure of the radiation exposure level, expressed in rads.}
\item{\code{bomb}}{a factor, indicating which bomb the subject survived (H = Hiroshima, N = Nagasaki).}
}
}
\description{
Count/Percentage of chromosome aberrations in atomic bombs survivors.
}
\details{
The data have been originally analyzed by Otake and Prentice (1984) and successively by Ascari and Migliorati (2021).
}
\references{
{
Ascari, R., Migliorati, S. (2021). A new regression model for overdispersed binomial data accounting for outliers and an excess of zeros. Statistics in Medicine, \bold{40}(17), 3895--3914. doi:10.1002/sim.9005 \cr
\cr
Otake, M., Prentice, R.L. (1984). The analysis of chromosomally aberrant cells based on beta-binomial distribution. Radiat Res. \bold{98}, 456--470. \cr
}
}
