% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\name{Bacteria}
\alias{Bacteria}
\title{Bacteria data}
\format{
A data frame containing 70 observations on the following 4 variables.
\describe{
\item{\code{y}}{the total number of parasitized eggs.}
\item{\code{n}}{the maximum number of eggs that female parasitoids could parasitized. It is fixed to 128 for all the observations.}
\item{\code{females}}{the number of female parasitoids.}
\item{\code{females_std}}{the standardized version of \code{females}.}
}
}
\source{
Demétrio et al., (2014). Models for overdispersed data in entomology.
}
\description{
Counts of eggs parasitized by female parasitoids.
}
\details{
The data have been originally analyzed by Demétrio et al (2014) and successively by Ascari and Migliorati (2021).
Data come from a completely randomized experiment with 10 replicates for each specification of number of females.
}
\references{
{
Ascari, R., Migliorati, S. (2021). A new regression model for overdispersed binomial data accounting for outliers and an excess of zeros. Statistics in Medicine, \bold{40}(17), 3895--3914. doi:10.1002/sim.9005 \cr
\cr
Demétrio, C.G.B., Hinde, J., Moral, R.A. (2014). Models for overdispersed data in entomology. Ecological Modelling Applied to Entomology. Entomology in Focus Switzerland: Springer International Publishing, 219--259. \cr
}
}
