% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.plot.R
\name{convergence.diag.internal}
\alias{convergence.diag.internal}
\title{internal function}
\usage{
convergence.diag.internal(
  posterior,
  diagnostics = "all",
  pars = NULL,
  additional.args = list(),
  aug = NULL,
  n.chain
)
}
\description{
internal function
}
\keyword{internal}
