% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.plot.R
\name{convergence.plot.internal}
\alias{convergence.plot.internal}
\title{internal function}
\usage{
convergence.plot.internal(
  posterior,
  file,
  plotfun,
  pars,
  point_est,
  prob,
  prob_outer,
  lags,
  warmup,
  width,
  height
)
}
\description{
internal function
}
\keyword{internal}
