% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\name{Election}
\alias{Election}
\title{Italian Election Results}
\format{
A data frame containing 232 observations on 13 variables.
\describe{
\item{\code{NVotes}}{the number of valid votes.}
\item{\code{FI}}{the percentage of votes got by `Forza Italia` party.}
\item{\code{FDI}}{the percentage of votes got by `Fratelli d'Italia` party.}
\item{\code{LEGA}}{the percentage of votes got by `Lega` party.}
\item{\code{LEU}}{the percentage of votes got by `Liberi e Uguali` party.}
\item{\code{M5S}}{the percentage of votes got by `Movimento 5 Stelle` party.}
\item{\code{PD}}{the percentage of votes got by `Partito Democratico` party.}
\item{\code{Other}}{the percentage of votes got by other parties, including blank ballots.}
\item{\code{AgeInd}}{the age index, defined as the ratio of the number of elderly persons (aged 65 and over) to the number of young persons (from 0 to 14), multiplied by 100.}
\item{\code{PopDens}}{the number of inhabitants per km2.}
\item{\code{ER}}{the employment rate, defined as the ratio of the number of employed persons (aged 15-64) to the number of persons (aged 15-64).}
\item{\code{Illiteracy}}{the illiteracy rate, defined as the ratio of the number of persons without a qualification (aged 15 and over) to the total number of persons aged 15 and over.}
\item{\code{Foreign}}{the number of foreigners per 1000 inhabitants.}
}
}
\source{
\href{http://www.interno.gov.it/it/speciali/2018-elections}{Italian Ministry of Interior’s webpage}.
}
\description{
Results of the Italian general election held on 4 March 2018 for six parties.
}
\details{
Data are collected on the 232 electoral districts into which the Italian territory is organized. Distribution of votes for Aosta constituency is not available. Distributions of votes are available on the Italian Ministry of Interior’s webpage whereas constituencies information have been obtained from 2011 Italian Census.
}
