% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexreg_Augmentation.R
\name{fit.model_2}
\alias{fit.model_2}
\title{internal function}
\usage{
fit.model_2(
  model.phi = model.phi,
  type = type,
  N = N,
  y = y,
  X = X,
  X0 = X0,
  X1 = X1,
  Z = Z,
  link_code_mu = link_code_mu,
  aug_code,
  link_prior_beta,
  hyperparam.beta,
  link_prior_omega0,
  hyperparam.omega0,
  link_prior_omega1,
  hyperparam.omega1,
  hyperparam.p,
  hyperparam.w,
  hyperparam.k,
  link_code_phi = link_code_phi,
  prior_code_phi = prior_code_phi,
  hyper_phi = hyper_phi,
  link_prior_psi,
  hyperparam.psi,
  n.iter,
  warmup.perc,
  n.chain,
  thin,
  verbose,
  ...
)
}
\description{
internal function
}
\keyword{internal}
