% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexreg_binom.R
\name{fit.model_binom}
\alias{fit.model_binom}
\title{internal function}
\usage{
fit.model_binom(
  model.theta = model.theta,
  type = type,
  N = N,
  n = n,
  y = y,
  X = X,
  Z = Z,
  link_code_mu = link_code_mu,
  link_prior_beta = link_prior_beta,
  hyperparam.beta = hyperparam.beta,
  link_code_theta = link_code_theta,
  hyper.theta.a = hyper.theta.a,
  hyper.theta.b = hyper.theta.b,
  hyperparam.p = hyperparam.p,
  hyperparam.w = hyperparam.w,
  link_prior_psi = link_prior_psi,
  hyperparam.psi = hyperparam.psi,
  n.iter,
  warmup.perc,
  n.chain,
  thin,
  verbose,
  ...
)
}
\description{
internal function
}
\keyword{internal}
