% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexReg-package.R
\docType{package}
\name{FlexReg-package}
\alias{FlexReg-package}
\alias{FlexReg}
\title{The 'FlexReg' package.}
\description{
The FlexReg package provides functions and methods to implement several types of regression models for bounded continuous responses (e.g., proportions and rates) and
bounded discrete responses (e.g., number of successes in n trials).
Inferential statistical analysis is dealt with by a Bayesian estimation procedure based on the Hamiltonian Monte Carlo (HMC) algorithm through the \code{rstan} package.
}
\references{
Ascari, R., Migliorati, S. (2021). A new regression model for overdispersed binomial data accounting for outliers and an excess of zeros. Statistics in Medicine, \bold{40}(17), 3895--3914. doi:10.1002/sim.9005 \cr
\cr
Ascari, R., Di Brisco, A. M., Migliorati, S., Ongaro, A. (2026). {FlexReg}: An {R} package for fitting a general class of mixture regression models with bounded responses in a {B}ayesian framework. Computational Statistics. doi:10.1007/s00180-026-01720-y \cr
\cr
Di Brisco, A. M., Migliorati, S. (2020). A new mixed-effects mixture model for constrained longitudinal data. Statistics in Medicine, \bold{39}(2), 129--145. doi:10.1002/sim.8406 \cr
\cr
Di Brisco, A. M., Migliorati, S., Ongaro, A. (2020). Robustness against outliers: A new variance inflated regression model for proportions. Statistical Modelling, \bold{20}(3), 274--309. doi:10.1177/1471082X18821213 \cr
\cr
Ferrari, S.L.P., Cribari-Neto, F. (2004). Beta Regression for Modeling Rates and Proportions. Journal of Applied Statistics, \bold{31}(7), 799--815. doi:10.1080/0266476042000214501\cr
\cr
Migliorati, S., Di Brisco, A. M., Ongaro, A. (2018). A New Regression Model for Bounded Responses. Bayesian Analysis, \bold{13}(3), 845--872. doi:10.1214/17-BA1079 \cr
\cr
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
\author{
\strong{Maintainer}: Roberto Ascari \email{roberto.ascari@unimib.it}

Authors:
\itemize{
  \item Agnese M. Di Brisco
  \item Sonia Migliorati
  \item Andrea Ongaro
}

}
\keyword{internal}
