% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{predict_lambda.chain}
\alias{predict_lambda.chain}
\title{predict_lambda.chain}
\usage{
predict_lambda.chain(posterior, mu.chain, newdata)
}
\description{
predict_lambda.chain
}
\keyword{internal}
