% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{predict_variance}
\alias{predict_variance}
\title{predict_variance}
\usage{
predict_variance(
  model,
  newdata,
  cluster,
  cluster.var = T,
  model.type,
  model.class,
  n
)
}
\description{
predict_variance
}
\keyword{internal}
