% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaries.R
\name{print.flexreg}
\alias{print.flexreg}
\title{Print Methods for flexreg Objects}
\usage{
\method{print}{flexreg}(x, ...)
}
\arguments{
\item{x}{an object of class \code{`flexreg`}.}

\item{...}{additional arguments. Currently not used.}
}
\description{
Print Methods for flexreg Objects
}
