% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\name{Reading}
\alias{Reading}
\title{Reading Skills data}
\format{
A data frame containing 44 observations on 3 variables.
\describe{
\item{\code{accuracy}}{a reading score.}
\item{\code{dyslexia}}{a factor indicating wheter the child is dyslexic.}
\item{\code{iq}}{a quantitative measure of the children's non verbal abilities.}
}
}
\source{
\href{https://CRAN.R-project.org/package=betareg}{betareg}.
}
\description{
Data for assessing the contribution of non-verbal IQ to children's reading skills in dyslexic and non-dyslexic children.
}
\details{
The data were originally analyzed by Pammer and Kevan (2004) and successively used by Smithson and Verkuilen (2006) and by Migliorati et al. (2018).
}
\references{
{
Smithson, M., and Verkuilen, J. (2006). A Better Lemon Squeezer? Maximum-Likelihood Regression with Beta-Distributed Dependent Variables. Psychological Methods, \bold{11}(7), 54–71. \cr
\cr
Cribari-Neto, F., and Zeileis, A. (2010). Beta Regression in R. Journal of Statistical Software, 34(2), 1–24. \cr
\cr
Migliorati, S., Di Brisco, A. M., Ongaro, A. (2018). A New Regression Model for Bounded Responses. Bayesian Analysis, \bold{13}(3), 845--872. doi:10.1214/17-BA1079 \cr
}
}
