% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\name{Stress}
\alias{Stress}
\title{Stress and anxiety data}
\format{
A data frame containing 166 observations on 2 variables.
\describe{
\item{\code{stress}}{ defined as rate. }
\item{\code{anxiety}}{defined as rate.}
}
}
\source{
Example 2 from Smithson and Verkuilen (2006).
}
\description{
Data for assessing the dependency between stress and anxiety in  nonclinical women in Townsville, Queensland, Australia.
}
\details{
Both variables are rates obtained as linear transformations from the Depression Anxiety Stress Scales which range from 0 to 42 (Lovibond & Lovibond, 1995). Additional details can be found in Example 2 from Smithson and Verkuilen (2006).
}
\references{
{
Smithson, M., and Verkuilen, J. (2006). A Better Lemon Squeezer? Maximum-Likelihood Regression with Beta-Distributed Dependent Variables. Psychological Methods, 11(7), 54–71. \cr
\cr Lovibond, P. F., & Lovibond, S. H. (1995). The structure of negative emotional states: Comparison of the Depression Anxiety Stress Scales (DASS) with the Beck Depression and Anxiety Inventories. Behaviour research and therapy, 33(3), 335-343.
}
}
