% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{var.fun}
\alias{var.fun}
\title{var.fun}
\usage{
var.fun(
  model.class,
  model.type,
  posterior,
  mu.chain,
  phi.chain,
  theta.chain,
  q0.chain,
  q1.chain,
  l1.chain,
  l2.chain,
  n
)
}
\description{
var.fun
}
\keyword{internal}
