% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexReg-package.R
\docType{package}
\name{FlexReg-package}
\alias{FlexReg-package}
\alias{FlexReg}
\title{The 'FlexReg' package.}
\description{
This package provides functions and methods to implement several types of regression models for bounded responses such as proportions and rates.
Inferential statistical analysis is dealt with by Bayesian estimation based on the Hamiltonian Monte Carlo (HMC) algorithm through the \code{`rstan`} package.
}
\references{
Migliorati, S., Di Brisco, A. M., Ongaro, A. (2018) A New Regression Model for Bounded Responses. Bayesian Analysis, \bold{13}(3), 845--872. doi:10.1214/17-BA1079 \cr
\cr
Di Brisco, A. M., Migliorati, S., Ongaro, A. (2020) Robustness against outliers: A new variance inflated regression model for proportions. Statistical Modelling, \bold{20}(3), 274--309. doi:10.1177/1471082X18821213 \cr
\cr
Ferrari, S.L.P., and Cribari-Neto, F. (2004). Beta Regression for Modeling Rates and Proportions. Journal of Applied Statistics, \bold{31}(7), 799--815. doi:10.1080/0266476042000214501\cr
\cr
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
