% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexreg_Augmentation.R
\name{flexreg}
\alias{flexreg}
\title{Flexible Regression Models for Proportions}
\usage{
flexreg(
  formula,
  zero.formula = NULL,
  one.formula = NULL,
  data,
  type = "FB",
  link.mu = "logit",
  prior.beta = "normal",
  hyperparam.beta = NULL,
  prior.omega0 = "normal",
  hyperparam.omega0 = NULL,
  prior.omega1 = "normal",
  hyperparam.omega1 = NULL,
  link.phi = NULL,
  prior.phi = NULL,
  hyperparam.phi = NULL,
  prior.psi = NULL,
  hyperparam.psi = NULL,
  n.iter = 5000,
  burnin.perc = 0.5,
  n.chain = 1,
  thin = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{`formula`}: a symbolic description of the mean model (\code{y ~ x}) or the mean and precision model (\code{y ~ x | z}) to be fitted.}

\item{zero.formula}{an object of class \code{`formula`}: a symbolic description of the zero augmented model to be fitted (see Details).}

\item{one.formula}{an object of class \code{`formula`}: a symbolic description of the one augmented model to be fitted (see Details).}

\item{data}{an optional data frame, list, or object that is coercible to a data frame through \code{base::as.data.frame} containing the variables in the model. If not found in data, the variables in formula are taken from the environment from which the function flexreg is called.}

\item{type}{a character specifying the type of regression model. Current options are \code{"FB"} (flexible beta, default), \code{"VIB"} (variance inflated beta), and \code{"Beta"}.}

\item{link.mu}{a character specifying the link function for the mean model (mu). Currently, \code{"logit"} (default), \code{"probit"}, \code{"cloglog"}, and \code{"loglog"} are supported.}

\item{prior.beta}{a character specifying the prior distribution for the  regression coefficients of the mean model, \code{beta}. Currently, \code{"normal"} (default) and \code{"cauchy"} are supported.}

\item{hyperparam.beta}{a positive numeric (vector of length 1) specifying the hyperprior standard deviation parameter for the prior distribution of \code{beta} regression coefficients. A value of 100 is suggested if the prior is \code{"normal"}, 2.5 if \code{"cauchy"}.}

\item{prior.omega0}{a character specifying the prior distribution for the  regression coefficients of the augmented model in zero, \code{omega0}. Currently, \code{"normal"} (default) and \code{"cauchy"} are supported.}

\item{hyperparam.omega0}{a positive numeric (vector of length 1) specifying the hyperprior standard deviation parameter for the prior distribution of \code{omega0} regression coefficients. A value of 100 is suggested if the prior is \code{"normal"}, 2.5 if \code{"cauchy"}.}

\item{prior.omega1}{a character specifying the prior distribution for the  regression coefficients of the augmented model in one, \code{omega1}. Currently, \code{"normal"} (default) and \code{"cauchy"} are supported.}

\item{hyperparam.omega1}{a positive numeric (vector of length 1) specifying the hyperprior standard deviation parameter for the prior distribution of \code{omega1} regression coefficients. A value of 100 is suggested if the prior is \code{"normal"}, 2.5 if \code{"cauchy"}.}

\item{link.phi}{a character specifying the link function for the precision model (phi). Currently, \code{"identity"} (default), \code{"log"}, and \code{"sqrt"} are supported.}

\item{prior.phi}{a character specifying the prior distribution for precision parameter \code{phi} if \cr \code{link.phi = "identity"}. Currently, \code{"gamma"} (default) and \code{"unif"} are supported.}

\item{hyperparam.phi}{a positive numeric (vector of length 1) specifying the hyperprior parameter for the prior distribution of \code{phi}. If the prior is \code{"gamma"}, the value identifies the gamma's shape and rate parameters (a value of 0.001 is suggested). If the prior is \code{"uniform"} the hyperparameter must be specified to define the upper limit of the support of \code{phi}.}

\item{prior.psi}{a character specifying the prior distribution for the regression coefficients of the precision model \code{psi} (not supported if \code{link.phi = "identity"}). Currently, \code{"normal"} (default) and \code{"cauchy"} are supported.}

\item{hyperparam.psi}{a positive numeric (vector of length 1) specifying the hyperprior standaerd deviation parameter for the prior distribution of \code{psi} regression coefficients. A value of 100 is suggested if the prior is \code{"normal"}, 2.5 if \code{"cauchy"}.}

\item{n.iter}{a positive integer specifying the number of iterations for each chain (including warmup). The default is 5000.}

\item{burnin.perc}{the percentage of iterations per chain to discard.}

\item{n.chain}{a positive integer specifying the number of Markov chains. The default is 1.}

\item{thin}{a positive integer specifying the period for saving samples. The default is 1.}

\item{verbose}{\code{TRUE} (default) or \code{FALSE}: flag indicating whether to print intermediate output.}

\item{...}{additional arguments for \code{rstan::sampling}.}
}
\value{
The \code{flexreg} function returns an object of class \code{`flexreg`}, i.e. a list with the following elements:
\item{\code{call}}{the function call.}
\item{\code{type}}{the type of regression model.}
\item{\code{formula}}{the overall formula.}
\item{\code{link.mu}}{a character specifing the link function in the mean model.}
\item{\code{link.phi}}{a character specifing the link function in the precision model.}
\item{\code{model}}{an object of class \code{`stanfit`} containing the fitted model.}
\item{\code{response}}{the response variable, assuming values in (0, 1).}
\item{\code{design.X}}{the design matrix for the mean model.}
\item{\code{design.Z}}{the design matrix for the precision model (if defined).}
\item{\code{design.X0}}{the design matrix for the augmented model in zero (if defined).}
\item{\code{design.X1}}{the design matrix for the augmented model in one (if defined).}
}
\description{
The function fits some flexible regression models for continuous bounded responses (e.g., proportions and rates) via a Bayesian approach to inference based on Hamiltonian Monte Carlo algorithm.
Available regression models are the flexible beta regression model (\code{type="FB"}, default), the variance inflated beta (\code{type="VIB"}),  the beta  (\code{type="Beta"}), and their augmented versions.
}
\details{
Let \eqn{\mu} be the mean of a random variable Y, whose distribution can be specified in the \code{type} argument.
The \code{flexreg} function links the parameter \eqn{\mu} to a linear predictor through a function  \eqn{g(\cdot)} specified in \code{link.mu}:
\deqn{g(\mu_i) = \bold{x}_i^t \bold{\beta},} where \eqn{\bold{\beta}} is the vector of regression coefficients for the mean model.
The prior distribution and the related hyperparameter of \eqn{\bold{\beta}} can be specified in \code{prior.beta} and \code{hyperparam.beta}.
By default, the precision parameter \eqn{\phi} is assumed to be constant.
The prior distribution and the related hyperparameter of \eqn{\phi} can be specified in \code{prior.phi} and \code{hyperparam.phi}.
It is possible to extend the model by linking \eqn{\phi} to an additional (possibly overlapping) set of covariates through a proper link
function \eqn{q(\cdot)}  specified in the \code{link.phi} argument:
\deqn{q(\phi_i) = \bold{z}_i^t \bold{\psi},} where \eqn{\bold{\psi}} is the vector of regression coefficients for the precision model.
The prior distribution and the related hyperparameter of \eqn{\bold{\psi}} can be specified in \code{prior.psi} and \code{hyperparam.psi}.
In \code{flexreg}, the regression model for the mean and, where appropriate, for the precision parameter can be specified in the
\code{formula} argument with a formula of type \code{y ~ x1 + x2 | z1 + z2} where covariates on the left of "|" are included in the regression model
for the mean and covariates on the right of "|" are included in the regression model for the precision.

If the second part is omitted, i.e., \code{y ~ x1 + x2}, the precision is assumed constant for each observation.

In presence of zero responses, one has to link the parameter \eqn{q_0} to an additional (possibly overlapping) set of covariates through a logit link function:
\deqn{g_0(q_{0i}) = \bold{x}_{0i}^t \bold{\omega_0},} where \eqn{\bold{\omega_0}} is the vector of regression coefficients for the augmented model in zero.
The prior distribution and the related hyperparameter of \eqn{\bold{\omega_0}} can be specified in \code{prior.omega0} and \code{hyperparam.omega0}.
In presence of one responses, one has to link the parameter \eqn{q_1} to an additional (possibly overlapping) set of covariates through a logit link function:
\deqn{g_1(q_{1i}) = \bold{x}_{1i}^t \bold{\omega_1},} where \eqn{\bold{\omega_1}} is the vector of regression coefficients for the augmented model in one.
The prior distribution and the related hyperparameter of \eqn{\bold{\omega_1}} can be specified in \code{prior.omega1} and \code{hyperparam.omega1}.
If both the augmented models in zero and one are specified, the link function is a bivariate logit.
In \code{flexreg}, the augmented models in zero and/or one can be specified in the
\code{zero.formula} and/or \code{one.formula} argument with a formula of type \code{ ~ x}.
Left hand side in \code{zero.formula} and \code{one.formula} can be omitted; if specified they have to be the same as left hand side in \code{formula}.
}
\examples{
\dontrun{
data("Reading")
FB <- flexreg(accuracy.adj ~ iq, data = Reading, type="FB")

# Regression model with one augmentation:
AFB1 <- flexreg(accuracy ~ dyslexia | iq + dyslexia + iq:dyslexia,
one.formula = ~ iq + dyslexia, data = Reading, type="FB")
}
}
\references{
{
Di Brisco, A. M., Migliorati, S. (2020). A new mixed-effects mixture model for constrained longitudinal data. Statistics in Medicine, \bold{39}(2), 129--145. doi:10.1002/sim.8406 \cr
\cr
Di Brisco, A. M., Migliorati, S., Ongaro, A. (2020). Robustness against outliers: A new variance inflated regression model for proportions. Statistical Modelling, \bold{20}(3), 274--309. doi:10.1177/1471082X18821213 \cr
\cr
 Ferrari, S.L.P., Cribari-Neto, F. (2004). Beta Regression for Modeling Rates and Proportions. Journal of Applied Statistics, \bold{31}(7), 799--815. doi:10.1080/0266476042000214501 \cr
\cr
Migliorati, S., Di Brisco, A. M., Ongaro, A. (2018) A New Regression Model for Bounded Responses. Bayesian Analysis, \bold{13}(3), 845--872. doi:10.1214/17-BA1079 \cr
}
}
